Upload Example 4.21

+++ Part 4 - Testing Express servers, user administration/blog/controllers/blogs.js
@@ -37,9 +37,24 @@ blogRouter.get('/', async (request, response) => {
     }
   })
 
-  blogRouter.delete('/:id', async (request, response) => {
-    await Blog.findByIdAndRemove(request.params.id)
-    response.status(204).end()
+  blogsRouter.delete('/:id', async (request, response, next) => {
+    const blog = await Blog.findById(request.params.id)
+    const decodedToken = jwt.verify(request.token, process.env.SECRET)
+    if (!request.token || !decodedToken.id) {
+      return response.status(401).json({ error: 'token missing or invalid' })
+    }
+    
+    try {
+      if (blog.user.toString() !== decodedToken.id.toString()) {
+        await Blog.findByIdAndRemove(request.params.id)
+        response.status(204).end()
+      } else {
+        return response.status(401).json({ error: 'you are not the author of the blog post' })
+      }
+    } catch (exception) {
+      response.status(400).end()
+      next(exception)
+    }
   })
 
   blogRouter.put('/:id', async (request, response) => {
